# framework/orchestrator/src/main/java/org/intueri/orchestrator/api/DetectorsApi.java
/**
 * Copyright (C) 2019 ${author} <speckij@posteo.net>
 * <p>
 * This software may be modified and distributed under the terms
 * of the MIT license. See the LICENSE file for details.
 */
package org.intueri.orchestrator.api;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import javax.validation.Valid;

public interface DetectorsApi {

    @RequestMapping(value = "/detectors",
            produces = {"application/json"},
            method = RequestMethod.GET)
    ResponseEntity<String> getDetectors();

    @RequestMapping(value = "/detectors/{id}",
            produces = {"application/json"},
            consumes = {"application/json"},
            method = RequestMethod.PUT)
    ResponseEntity<Void> detectorsIdPut(@Valid @RequestBody String body, @PathVariable("id") String id);

}